fix: support decimal scores in rank progress API and energy service
- Updated backend ranks API to accept float final_score instead of int
- Fixed energy service to handle user gender validation with proper fallbacks
- Modified score API to use correct parameters for energy service calls
- Enhanced result screen with debug logging for API troubleshooting
- Maintained decimal precision throughout frontend while ensuring backend compatibility

The changes enable proper handling of decimal scores for rank calculations
while fixing gender validation issues in energy service computations.

+++ backend/app/api/v1/ranks.py
@@ -141,7 +141,7 @@ async def rank_icon(user_id: str):
 @router.get("/get_rank_progress", response_model=dict)
 async def get_rank_progress(
     scenario_id: str, 
-    final_score: int,  # This should already be the calculated score_value from the frontend
+    final_score: float,
     user_weight: float, 
     user_gender: str = "male"
 ):
